Added the ability to lazy-load or force manual loading of GeoName ancestry, setting the defaults for all queries to lazy loading. This should improve performance, bringing it back to CLAVIN 1.x levels for simple lookups where ancestry is not desired. In order to achieve this, GeoName was refactored to an interface to allow a wrapper object that provides the lazy load functionality. The Gazetteer interface has been updated to provide methods for loading ancestry for one or more GeoNames and to specify an AncestryMode when loading a GeoName by ID.
