+++ src/main/java/com/bericotech/clavin/gazetteer/CountryCode.java
@@ -1,31 +1,31 @@
 package com.bericotech.clavin.gazetteer;
 
 /*#####################################################################
- * 
+ *
  * CLAVIN (Cartographic Location And Vicinity INdexer)
  * ---------------------------------------------------
- * 
+ *
  * Copyright (C) 2012-2013 Berico Technologies
  * http://clavin.bericotechnologies.com
- * 
+ *
  * ====================================================================
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
  * implied. See the License for the specific language governing
  * permissions and limitations under the License.
- * 
+ *
  * ====================================================================
- * 
+ *
  * CountryCode.java
- * 
+ *
  *###################################################################*/
 
 /**
@@ -285,19 +285,19 @@ public enum CountryCode {
     ZW("Zimbabwe", 878675),
     CS("Serbia and Montenegro", GeoName.OUT_OF_BOUNDS), // no longer exists
     AN("Netherlands Antilles", GeoName.OUT_OF_BOUNDS), // no longer exists
-    
+
     // manually added for locations not assigned to a specific country
     NULL("No Man's Land", GeoName.OUT_OF_BOUNDS);
-    
+
     // country name
     public final String name;
-    
+
     // unique identifier
     public final int geonameID;
-    
+
     /**
      * Constructor for {@link CountryCode} enum type.
-     * 
+     *
      * @param name          country name
      * @param geonameID     unique identifier
      */
@@ -305,5 +305,5 @@ public enum CountryCode {
         this.name = name;
         this.geonameID = geonameID;
     }
-    
+
 }
