+++ src/test/resources/logback-test.xml
@@ -7,8 +7,10 @@
       <pattern>%d{HH:mm:ss.SSS} [%thread] %-5level - %msg%n</pattern>
     </encoder>
   </appender>
-   
-  <root level="ALL">
+
+  <root level="INFO">
     <appender-ref ref="STDOUT" />
   </root>
+
+  <!--<logger name="com.bericotech.clavin" level="debug"/>-->
 </configuration>
